+++ modcore/__init__.py
@@ -14,7 +14,7 @@ from .mod import Module
 from .lifecycle import LifeCycle
 from .log import DEBUG, INFO, NOTSET, logger
 
-VERSION = "v0.0.12"
+VERSION = "v0.0.13a"
 
 print( "-"*41 )
 print( "mpy modcore" )