+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## next version v0.0.12
 
 - variable `$id$` as part of ssid name in `softap.cfg` will be expanded to machine unique id
+- intr Counter rework
 -